Split the app module into focused submodules
app.rs had grown to 2,600 lines carrying five concerns at once: core
state and event folding, the detail-pane content, keyboard input, mouse
input with the scrollbar arithmetic, and every action's implementation.
Navigating it meant scrolling past all of them.

  app/mod.rs      state, startup, sidebar, background loads, events
  app/detail.rs   what the detail pane contains (the content walk)
  app/keys.rs     keyboard input and the rename fields
  app/mouse.rs    hit testing, hover, wheel routing, scrollbar drags
  app/actions.rs  what activating a control does, and the spawned flows

Pure code motion — no behaviour change; the public surface is re-
exported from app so no caller moved.
