+++ src/app/mod.rs
@@ -1,21 +1,25 @@
-//! Application state and behaviour: focus, key handling, and background work.
+//! Application state and the event loop's core: startup, the sidebar, the
+//! background loads a selection kicks off, and event folding.
+//!
+//! The input handling lives beside it in focused modules: [`keys`] for the
+//! keyboard, [`mouse`] for hit testing and scrolling, [`actions`] for what a
+//! control does when activated, and [`detail`] for what the detail pane
+//! contains.
 
+mod actions;
 pub mod detail;
+mod keys;
+mod mouse;
 
 pub use detail::{Action, DetailItem, Field};
+pub use mouse::{Hit, HitTarget, ModalClick, ScrollbarGeom};
 
-use crate::event::{AppEvent, BranchTarget, DetailMeta, EventTx, Severity};
-use crate::modal::{
-    AddRepositoryModal, AddWorktreeModal, ConfirmAction, ConfirmModal, CreateFromIssueModal, Modal,
-    ModalEffect, ModalOutcome, ModalResult, SettingsModal,
-};
-use crate::models::{
-    ClaudeSession, CustomClaudeButton, GitHubIssue, Repository, Settings, Worktree,
-};
+use crate::event::{AppEvent, DetailMeta, EventTx, Severity};
+use crate::modal::Modal;
+use crate::models::{ClaudeSession, GitHubIssue, Settings};
 use crate::services::{claude_session, git, github, settings as settings_service, tmux};
 use crate::state::AppState;
 use crate::ui::widgets::TextInput;
-use std::path::{Path, PathBuf};
 use std::time::{Duration, Instant};
 use uuid::Uuid;
 
@@ -26,84 +30,9 @@ pub const SESSION_LIMIT: usize = 5;
 pub const ISSUE_LIMIT: usize = 10;
 /// Lines the detail pane moves per wheel notch, matching Textual's
 /// `scroll_sensitivity_y` of 3.
-const SCROLL_STEP: isize = 3;
+pub(super) const SCROLL_STEP: isize = 3;
 /// Rows `PageUp` / `PageDown` move.
-const PAGE_STEP: isize = 10;
-
-/// Pointer motion, held button or not. Either may well leave the frame
-/// identical, so both are left to the handler to mark dirty when it acts.
-fn is_pointer_motion(event: &AppEvent) -> bool {
-    use ratatui::crossterm::event::{Event, MouseEventKind};
-    matches!(
-        event,
-        AppEvent::Term(Event::Mouse(mouse))
-            if matches!(mouse.kind, MouseEventKind::Moved | MouseEventKind::Drag(_))
-    )
-}
-
-/// Where the detail pane's scrollbar was drawn, and the arithmetic that maps
-/// between a thumb position and a scroll offset.
-///
-/// Kept in one place because the renderer and the drag handler have to agree
-/// exactly: a thumb that renders one row off from where the drag thinks it is
-/// feels like the bar slipping out from under the pointer.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct ScrollbarGeom {
-    /// The full track, including the part the thumb is not covering.
-    pub track: ratatui::layout::Rect,
-    pub thumb: u16,
-    /// Rows the top of the thumb can travel.
-    pub travel: u16,
-    pub max_offset: u16,
-}
-
-impl ScrollbarGeom {
-    /// `None` when the content fits, which is also when no bar is drawn.
-    pub fn new(track: ratatui::layout::Rect, total: u16) -> Option<Self> {
-        let height = track.height;
-        if height == 0 || total <= height {
-            return None;
-        }
-        // At least one cell, so the thumb never vanishes on very long content.
-        let thumb = ((u32::from(height) * u32::from(height)) / u32::from(total)).max(1) as u16;
-        Some(Self {
-            track,
-            thumb,
-            travel: height.saturating_sub(thumb),
-            max_offset: total - height,
-        })
-    }
-
-    /// Rounds to nearest rather than truncating. There are usually far more
-    /// scroll positions than track rows, so truncating leaves the thumb a row
-    /// behind the offset a drag just produced — the bar visibly lags the pointer
-    /// and never quite reaches the bottom. Rounding makes this the exact inverse
-    /// of [`Self::offset_at`] for every row of the track.
-    pub fn thumb_top(&self, offset: u16) -> u16 {
-        if self.max_offset == 0 {
-            return 0;
-        }
-        let max = u32::from(self.max_offset);
-        let scaled = u32::from(offset.min(self.max_offset)) * u32::from(self.travel);
-        (((scaled + max / 2) / max) as u16).min(self.travel)
-    }
-
-    pub fn offset_at(&self, thumb_top: u16) -> u16 {
-        if self.travel == 0 {
-            return 0;
-        }
-        ((u32::from(thumb_top.min(self.travel)) * u32::from(self.max_offset))
-            / u32::from(self.travel)) as u16
-    }
-}
-
-/// Is this cell inside that rectangle?
-fn contains(rect: ratatui::layout::Rect, column: u16, row: u16) -> bool {
-    column >= rect.x
-        && column < rect.x.saturating_add(rect.width)
-        && row >= rect.y
-        && row < rect.y.saturating_add(rect.height)
-}
+pub(super) const PAGE_STEP: isize = 10;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum Focus {
@@ -154,51 +83,6 @@ pub struct Notification {
     pub text: String,
     pub severity: Severity,
     pub created: Instant,
-}
-
-/// Something a mouse click can land on.
-///
-/// Immediate mode keeps no widget tree, so there is nothing to ask "what is at
-/// this cell?". Each frame the renderers record the rectangle of every clickable
-/// thing, and clicks are resolved against that list.
-/// What clicking a modal control should do.
-///
-/// Not every control activates: a click that focuses a text field must not also
-/// submit the modal, and clicking `◂ value ▸` should advance the value rather
-/// than accept the dialog.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ModalClick {
-    /// Buttons and checkboxes: focus, then activate.
-    Activate,
-    /// Text inputs: focus only.
-    Focus,
-    /// A `◂ value ▸` cycle: focus, then advance one step.
-    Cycle,
-    /// A row of a list: focus the list and select that row.
-    Row(usize),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum HitTarget {
-    SidebarRow(usize),
-    /// The `▾`/`▸` twisty on a repository row, which folds it away without
-    /// changing the selection.
-    SidebarToggle(usize),
-    DetailItem(usize),
-    /// A key in the footer bar. Clicking one is the same as pressing it, so the
-    /// character is carried rather than a resolved action.
-    FooterKey(char),
-    /// Focus index within the modal on top of the stack, and what a click does.
-    ModalControl {
-        index: usize,
-        click: ModalClick,
-    },
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct Hit {
-    pub rect: ratatui::layout::Rect,
-    pub target: HitTarget,
 }
 
 pub struct App {
@@ -586,7 +470,7 @@ impl App {
         // repaint some other event in the same drain already earned. The tick
         // decides for itself in `on_tick`: at ten a second, repainting an idle
         // frame on every one keeps the terminal permanently busy for nothing.
-        if !is_pointer_motion(&event) && !matches!(event, AppEvent::Tick) {
+        if !mouse::is_pointer_motion(&event) && !matches!(event, AppEvent::Tick) {
             self.redraw = true;
         }
         match event {
@@ -684,1070 +568,12 @@ impl App {
         }
         handled
     }
-
-    // -------------------------------------------------------------------- mouse
-
-    /// Drop the previous frame's clickable regions. Called at the start of draw.
-    pub fn clear_hits(&mut self) {
-        self.hits.clear();
-    }
-
-    pub fn push_hit(&mut self, rect: ratatui::layout::Rect, target: HitTarget) {
-        self.hits.push(Hit { rect, target });
-    }
-
-    /// Topmost target at a cell. Later hits win, so a modal drawn over the panes
-    /// takes the click rather than whatever it covers.
-    pub fn hit_at(&self, column: u16, row: u16) -> Option<HitTarget> {
-        self.hits
-            .iter()
-            .rev()
-            .find(|hit| {
-                let r = hit.rect;
-                column >= r.x && column < r.x + r.width && row >= r.y && row < r.y + r.height
-            })
-            .map(|hit| hit.target)
-    }
-
-    pub fn handle_mouse(&mut self, mouse: ratatui::crossterm::event::MouseEvent) {
-        use ratatui::crossterm::event::{MouseButton, MouseEventKind};
-
-        match mouse.kind {
-            MouseEventKind::ScrollDown => {
-                self.scroll_at(mouse.column, mouse.row, 1);
-                return;
-            }
-            MouseEventKind::ScrollUp => {
-                self.scroll_at(mouse.column, mouse.row, -1);
-                return;
-            }
-            // Hover. The pointer crossing cells inside one control changes
-            // nothing on screen, so only a change of target costs a repaint.
-            MouseEventKind::Moved => {
-                let target = self.hit_at(mouse.column, mouse.row);
-                if target != self.hovered {
-                    self.hovered = target;
-                    self.redraw = true;
-                }
-                return;
-            }
-            // Dragging the scrollbar. The pointer routinely leaves the track
-            // mid-drag, so this is not gated on where it currently is — only on
-            // a drag having started there.
-            MouseEventKind::Drag(MouseButton::Left) => {
-                self.drag_scrollbar(mouse.row);
-                return;
-            }
-            // The release carries a position too, and it is often past the last
-            // drag report — a quick flick ends with the button coming up several
-            // rows beyond where the last motion landed. Applying it before
-            // letting go means the thumb finishes where the pointer did.
-            MouseEventKind::Up(MouseButton::Left) => {
-                self.drag_scrollbar(mouse.row);
-                self.scroll_drag = None;
-                return;
-            }
-            MouseEventKind::Down(MouseButton::Left) => {
-                if self.grab_scrollbar(mouse.column, mouse.row) {
-                    return;
-                }
-            }
-            _ => return,
-        }
-
-        let Some(target) = self.hit_at(mouse.column, mouse.row) else {
-            return;
-        };
-
-        // A modal owns every click while it is open, including the ones that
-        // land on the panes behind it.
-        if !self.modals.is_empty() {
-            if let HitTarget::ModalControl { index, click } = target {
-                self.click_modal_control(index, click);
-            }
-            return;
-        }
-
-        match target {
-            HitTarget::SidebarRow(index) => {
-                self.focus = Focus::Sidebar;
-                self.sidebar_index = index;
-                self.select_current_row();
-            }
-            HitTarget::SidebarToggle(index) => self.toggle_collapsed(index),
-            HitTarget::DetailItem(index) => {
-                self.focus = Focus::Detail;
-                self.detail_index = index;
-                self.detail_follow_focus = true;
-                if let Some(DetailItem::Action(action)) = self.detail_items().get(index).cloned() {
-                    self.run_action(action);
-                }
-            }
-            HitTarget::FooterKey(key) => {
-                self.handle_key(ratatui::crossterm::event::KeyEvent::new(
-                    ratatui::crossterm::event::KeyCode::Char(key),
-                    ratatui::crossterm::event::KeyModifiers::NONE,
-                ));
-            }
-            HitTarget::ModalControl { .. } => {}
-        }
-    }
-
-    /// Fold a repository row shut, or open it again.
-    ///
-    /// Folding is a view operation and leaves the selection alone — unless the
-    /// selected worktree is the thing being hidden. Keeping it would leave the
-    /// detail pane describing a row that is no longer on screen, with the
-    /// sidebar cursor pointing at whatever slid into its place, so the selection
-    /// falls back to the repository that swallowed it.
-    pub fn toggle_collapsed(&mut self, index: usize) {
-        let Some(SidebarRow::Repository { id, .. }) = self.rows.get(index) else {
-            return;
-        };
-        let id = *id;
-        let collapsing = self.collapsed.insert(id);
-        if !collapsing {
-            self.collapsed.remove(&id);
-        }
-
-        let hides_selection = collapsing
-            && self.state.selection.repository_id == Some(id)
-            && self.state.selection.worktree_id.is_some();
-        if hides_selection {
-            self.state.select_repository(id);
-            self.rebuild_rows();
-            self.sync_sidebar_index();
-            self.reload_detail();
-            return;
-        }
-
-        self.rebuild_rows();
-        self.sync_sidebar_index();
-    }
-
-    /// Route the wheel to the pane the pointer is over, not the focused one.
-    ///
-    /// The pointer is the thing the user is aiming with; Textual scrolled the
-    /// widget under it, and scrolling a pane the mouse is nowhere near is the
-    /// behaviour this replaces.
-    fn scroll_at(&mut self, column: u16, row: u16, delta: isize) {
-        if contains(self.sidebar_rect, column, row) {
-            self.scroll_sidebar(delta);
-        } else if contains(self.detail_rect, column, row) {
-            self.scroll_detail(delta);
-        }
-    }
-
-    /// Start a scrollbar drag, if the press landed on the bar.
-    ///
-    /// Pressing the thumb picks it up where it was grabbed. Pressing the bare
-    /// track pages the thumb to the pointer first and then behaves as if it had
-    /// been grabbed by its middle, so the press and any drag that follows are
-    /// one continuous gesture rather than a jump and then a second one.
-    ///
-    /// Focus is left alone on purpose: scrolling something is not the same as
-    /// aiming at it, and stealing focus here would move the keyboard cursor into
-    /// the pane every time the user only wanted to read further down.
-    fn grab_scrollbar(&mut self, column: u16, row: u16) -> bool {
-        let Some(geom) = self.scrollbar else {
-            return false;
-        };
-        if !contains(geom.track, column, row) {
-            return false;
-        }
-
-        let local = row - geom.track.y;
-        let top = geom.thumb_top(self.detail_scroll);
-        let grab = if local >= top && local < top.saturating_add(geom.thumb) {
-            local - top
-        } else {
-            let middle = geom.thumb / 2;
-            self.detail_scroll = geom.offset_at(local.saturating_sub(middle));
-            middle.min(local)
-        };
-
-        self.scroll_drag = Some(grab);
-        true
-    }
-
-    /// Continue a scrollbar drag. Silent unless a drag is actually in progress,
-    /// so a drag begun anywhere else cannot move the pane.
-    fn drag_scrollbar(&mut self, row: u16) {
-        let (Some(geom), Some(grab)) = (self.scrollbar, self.scroll_drag) else {
-            return;
-        };
-        // Above the track this saturates to the top; below it, clamping to the
-        // last row and letting `offset_at` cap the result pins it to the bottom.
-        let local = row
-            .saturating_sub(geom.track.y)
-            .min(geom.track.height.saturating_sub(1));
-        let next = geom.offset_at(local.saturating_sub(grab));
-        if next != self.detail_scroll {
-            self.detail_scroll = next;
-            self.redraw = true;
-        }
-    }
-
-    /// Scroll the detail pane's viewport. Deliberately independent of the focus
-    /// ring: reading something is not the same as aiming at it.
-    fn scroll_detail(&mut self, delta: isize) {
-        let next = (self.detail_scroll as isize + delta * SCROLL_STEP).max(0);
-        // The upper bound depends on content height, which only the renderer
-        // knows; it clamps there and writes the value back.
-        self.detail_scroll = next as u16;
-    }
-
-    fn scroll_sidebar(&mut self, delta: isize) {
-        self.move_sidebar(delta);
-    }
-
-    /// Apply a click to a modal control: always focus it, then do whatever that
-    /// kind of control does when driven from the keyboard.
-    fn click_modal_control(&mut self, index: usize, click: ModalClick) {
-        use ratatui::crossterm::event::KeyCode;
-
-        if let Some(modal) = self.modals.last_mut() {
-            modal.set_focus(index);
-            if let ModalClick::Row(row) = click {
-                modal.set_row(row);
-            }
-        }
-        match click {
-            ModalClick::Activate => self.send_modal_key(KeyCode::Enter),
-            ModalClick::Cycle => self.send_modal_key(KeyCode::Right),
-            // Focusing a field must not submit the dialog, and selecting a row
-            // is already done above.
-            ModalClick::Focus | ModalClick::Row(_) => {}
-        }
-    }
-
-    fn send_modal_key(&mut self, code: ratatui::crossterm::event::KeyCode) {
-        use ratatui::crossterm::event::{KeyEvent, KeyEventKind, KeyEventState, KeyModifiers};
-        self.handle_modal_key(KeyEvent {
-            code,
-            modifiers: KeyModifiers::NONE,
-            kind: KeyEventKind::Press,
-            state: KeyEventState::NONE,
-        });
-    }
-
-    // --------------------------------------------------------------------- keys
-
-    fn handle_key(&mut self, key: ratatui::crossterm::event::KeyEvent) {
-        use ratatui::crossterm::event::{KeyCode, KeyModifiers};
-
-        if key.modifiers.contains(KeyModifiers::CONTROL) && key.code == KeyCode::Char('c') {
-            self.should_quit = true;
-            return;
-        }
-
-        if !self.modals.is_empty() {
-            self.handle_modal_key(key);
-            return;
-        }
-
-        // A focused text field owns every printable key.
-        if self.focus == Focus::Detail
-            && let Some(DetailItem::Field(field)) =
-                self.detail_items().get(self.detail_index).cloned()
-            && self.handle_field_key(field, key)
-        {
-            return;
-        }
-
-        match key.code {
-            KeyCode::Tab => {
-                self.focus = match self.focus {
-                    Focus::Sidebar => Focus::Detail,
-                    Focus::Detail => Focus::Sidebar,
-                };
-                // Entering the pane should show where the cursor landed.
-                self.detail_follow_focus = true;
-                return;
-            }
-            KeyCode::Up => {
-                match self.focus {
-                    Focus::Sidebar => self.move_sidebar(-1),
-                    Focus::Detail => {
-                        self.detail_index = self.detail_index.saturating_sub(1);
-                        self.detail_follow_focus = true;
-                    }
-                }
-                return;
-            }
-            KeyCode::Down => {
-                match self.focus {
-                    Focus::Sidebar => self.move_sidebar(1),
-                    Focus::Detail => {
-                        let len = self.detail_items().len();
-                        if len > 0 {
-                            self.detail_index = (self.detail_index + 1).min(len - 1);
-                        }
-                        self.detail_follow_focus = true;
-                    }
-                }
-                return;
-            }
-            // Textual bound these to the viewport rather than the focus ring, so
-            // in the detail pane they page without moving the cursor.
-            KeyCode::PageUp => {
-                match self.focus {
-                    Focus::Sidebar => self.move_sidebar(-PAGE_STEP),
-                    Focus::Detail => self.scroll_detail(-PAGE_STEP),
-                }
-                return;
-            }
-            KeyCode::PageDown => {
-                match self.focus {
-                    Focus::Sidebar => self.move_sidebar(PAGE_STEP),
-                    Focus::Detail => self.scroll_detail(PAGE_STEP),
-                }
-                return;
-            }
-            KeyCode::Enter => {
-                match self.focus {
-                    Focus::Sidebar => self.select_current_row(),
-                    Focus::Detail => {
-                        if let Some(DetailItem::Action(action)) =
-                            self.detail_items().get(self.detail_index).cloned()
-                        {
-                            self.run_action(action);
-                        }
-                    }
-                }
-                return;
-            }
-            _ => {}
-        }
-
-        match key.code {
-            KeyCode::Char('q') => self.should_quit = true,
-            KeyCode::Char('a') => self
-                .modals
-                .push(Modal::AddRepository(AddRepositoryModal::new())),
-            KeyCode::Char('w') => self.action_add_worktree(),
-            KeyCode::Char('e') => self.run_action(Action::Editor),
-            KeyCode::Char('t') => self.run_action(Action::Terminal),
-            KeyCode::Char('o') => self.run_action(Action::Files),
-            KeyCode::Char('n') => self.run_action(Action::ClaudeNew),
-            KeyCode::Char('y') => self.run_action(Action::ClaudeYolo),
-            KeyCode::Char('h') => self.action_toggle_archive(),
-            KeyCode::Char('d') => self.action_delete(),
-            KeyCode::Char('s') => self
-                .modals
-                .push(Modal::Settings(Box::new(SettingsModal::new(
-                    &self.settings,
-                )))),
-            KeyCode::Char('r') => {
-                self.rebuild_rows();
-                self.reload_detail();
-            }
-            KeyCode::Char('A') => {
-                self.state.show_archived = !self.state.show_archived;
-                self.rebuild_rows();
-            }
-            KeyCode::Char('?') => self.notify(
-                "a: Add Repo | w: Add Worktree | e: Editor | t: Terminal | \
-                 n: Claude | h: Archive | d: Delete | s: Settings | q: Quit",
-                Severity::Information,
-            ),
-            _ => {}
-        }
-    }
-
-    /// Route a key to a focused rename field. Returns true when consumed.
-    fn handle_field_key(&mut self, field: Field, key: ratatui::crossterm::event::KeyEvent) -> bool {
-        use ratatui::crossterm::event::{KeyCode, KeyModifiers};
-
-        if matches!(
-            key.code,
-            KeyCode::Tab | KeyCode::BackTab | KeyCode::Up | KeyCode::Down
-        ) {
-            return false;
-        }
-        if key.code == KeyCode::Enter {
-            self.submit_rename(field);
-            return true;
-        }
-
-        let input = match field {
-            Field::WorktreeName => &mut self.name_input,
-            Field::BranchName => &mut self.branch_input,
-        };
-        match key.code {
-            KeyCode::Char('u') if key.modifiers.contains(KeyModifiers::CONTROL) => {
-                input.kill_to_start();
-                true
-            }
-            KeyCode::Char(c) if !key.modifiers.contains(KeyModifiers::CONTROL) => {
-                input.insert(c);
-                true
-            }
-            KeyCode::Backspace => {
-                input.backspace();
-                true
-            }
-            KeyCode::Delete => {
-                input.delete();
-                true
-            }
-            KeyCode::Left => {
-                input.move_left();
-                true
-            }
-            KeyCode::Right => {
-                input.move_right();
-                true
-            }
-            KeyCode::Home => {
-                input.move_home();
-                true
-            }
-            KeyCode::End => {
-                input.move_end();
-                true
-            }
-            // A focused field swallows every printable key, so Escape has to
-            // hand focus back as well as undo the edit — otherwise the global
-            // hotkeys are unreachable from here.
-            KeyCode::Esc => {
-                self.reset_rename_fields();
-                self.focus = Focus::Sidebar;
-                true
-            }
-            _ => false,
-        }
-    }
-
-    fn reset_rename_fields(&mut self) {
-        if let Some((_repo, worktree)) = self.state.selected_worktree() {
-            self.name_input = TextInput::new(worktree.name.clone());
-            self.branch_input = TextInput::new(worktree.branch.clone());
-        }
-    }
-
-    // ------------------------------------------------------------------- modals
-
-    fn handle_modal_key(&mut self, key: ratatui::crossterm::event::KeyEvent) {
-        let Some(modal) = self.modals.last_mut() else {
-            return;
-        };
-        match modal.handle_key(key) {
-            ModalOutcome::None => {}
-            ModalOutcome::Close => {
-                self.modals.pop();
-            }
-            ModalOutcome::Push(child) => self.modals.push(*child),
-            ModalOutcome::Effect(effect) => self.run_modal_effect(effect),
-            ModalOutcome::Submit(result) => {
-                self.modals.pop();
-                if let Some(parent) = self.modals.last_mut() {
-                    parent.receive_child(&result);
-                }
-                self.apply_modal_result(result);
-            }
-        }
-    }
-
-    fn run_modal_effect(&mut self, effect: ModalEffect) {
-        match effect {
-            ModalEffect::Fetch(repo_path) => {
-                let repo_id = self.state.selection.repository_id.unwrap_or_default();
-                let tx = self.tx.clone();
-                tokio::spawn(async move {
-                    if let Err(error) = git::fetch(&repo_path).await {
-                        tx.send(AppEvent::FetchFailed(error.to_string()));
-                        return;
-                    }
-                    let branches = git::list_branches(&repo_path, true)
-                        .await
-                        .unwrap_or_default();
-                    let remotes = git::list_remotes(&repo_path).await.unwrap_or_default();
-                    let current_branch = git::get_current_branch(&repo_path)
-                        .await
-                        .unwrap_or_else(|_| "main".to_string());
-                    tx.send(AppEvent::Branches {
-                        repo_id,
-                        branches,
-                        remotes,
-                        current_branch,
-                        target: BranchTarget::RefetchOpenModal,
-                    });
-                });
-            }
-        }
-    }
-
-    fn apply_modal_result(&mut self, result: ModalResult) {
-        match result {
-            ModalResult::RepositoryAdded {
-                path,
-                import_worktrees,
-            } => self.add_repository(path, import_worktrees),
-            ModalResult::WorktreeCreated {
-                repo_id,
-                name,
-                branch,
-                new_branch,
-            } => self.create_worktree(repo_id, name, branch, new_branch, None, false),
-            ModalResult::WorktreeFromIssue {
-                repo_id,
-                name,
-                branch,
-                pull_first,
-                base_branch,
-            } => self.create_worktree(repo_id, name, branch, true, base_branch, pull_first),
-            ModalResult::SettingsSaved(settings) => {
-                self.settings = *settings;
-                if let Err(error) = settings_service::save_settings(&self.settings) {
-                    self.notify(format!("Could not save settings: {error}"), Severity::Error);
-                } else {
-                    self.notify("Settings saved", Severity::Information);
-                }
-                self.detail_index = 0;
-            }
-            ModalResult::CustomButtonsSaved(_) | ModalResult::CustomButtonSaved(_) => {}
-            ModalResult::Confirmed(action) => self.apply_confirmed(action),
-        }
-    }
-
-    fn apply_confirmed(&mut self, action: ConfirmAction) {
-        match action {
-            ConfirmAction::RemoveRepository(repo_id) => {
-                self.state.remove_repository(repo_id);
-                self.rebuild_rows();
-                self.sync_sidebar_index();
-                self.reload_detail();
-            }
-            ConfirmAction::DeleteWorktree(worktree_id) => {
-                let Some((repo, worktree)) = self.state.find_worktree(worktree_id) else {
-                    return;
-                };
-                let repo_path = repo.source_path.clone();
-                let worktree_path = worktree.path.clone();
-                let tx = self.tx.clone();
-                tokio::spawn(async move {
-                    // A failed git removal must not strand the entry in state.
-                    let _ = git::remove_worktree(&repo_path, &worktree_path).await;
-                    tx.send(AppEvent::ReloadDetail);
-                });
-                self.state.remove_worktree(worktree_id);
-                self.rebuild_rows();
-                self.sync_sidebar_index();
-                self.reload_detail();
-            }
-        }
-    }
-
-    fn on_branches(
-        &mut self,
-        repo_id: Uuid,
-        branches: Vec<String>,
-        remotes: Vec<String>,
-        current_branch: String,
-        target: BranchTarget,
-    ) {
-        match target {
-            BranchTarget::RefetchOpenModal => {
-                if let Some(Modal::CreateFromIssue(modal)) = self.modals.last_mut() {
-                    modal.update_branches(branches, remotes, &current_branch);
-                }
-            }
-            BranchTarget::AddWorktree => {
-                let Some(repo) = self.state.find_repository(repo_id) else {
-                    return;
-                };
-                self.modals
-                    .push(Modal::AddWorktree(Box::new(AddWorktreeModal::new(
-                        repo,
-                        branches,
-                        remotes,
-                        settings_service::get_forest_path(),
-                        self.settings.branch_prefix.clone(),
-                    ))));
-            }
-            BranchTarget::CreateFromIssue => {
-                let Some(index) = self.pending_issue.take() else {
-                    return;
-                };
-                let Some(issue) = self.issues.as_ref().and_then(|i| i.get(index)).cloned() else {
-                    return;
-                };
-                let Some(repo) = self.state.find_repository(repo_id) else {
-                    return;
-                };
-                self.modals
-                    .push(Modal::CreateFromIssue(Box::new(CreateFromIssueModal::new(
-                        repo,
-                        &issue,
-                        branches,
-                        remotes,
-                        settings_service::get_forest_path(),
-                        &self.settings.branch_prefix,
-                        &current_branch,
-                    ))));
-            }
-        }
-    }
-
-    // ------------------------------------------------------------------ actions
-
-    fn load_branches_then(&mut self, repo_id: Uuid, target: BranchTarget) {
-        let Some(repo) = self.state.find_repository(repo_id) else {
-            return;
-        };
-        let path = repo.source_path.clone();
-        let tx = self.tx.clone();
-        tokio::spawn(async move {
-            let branches = git::list_branches(&path, true).await.unwrap_or_default();
-            let remotes = git::list_remotes(&path).await.unwrap_or_default();
-            let current_branch = git::get_current_branch(&path)
-                .await
-                .unwrap_or_else(|_| "main".to_string());
-            tx.send(AppEvent::Branches {
-                repo_id,
-                branches,
-                remotes,
-                current_branch,
-                target,
-            });
-        });
-    }
-
-    fn action_add_worktree(&mut self) {
-        match self.state.selection.repository_id {
-            Some(repo_id) => self.load_branches_then(repo_id, BranchTarget::AddWorktree),
-            None => self.notify("Select a repository first", Severity::Warning),
-        }
-    }
-
-    fn action_toggle_archive(&mut self) {
-        let Some(worktree_id) = self.state.selection.worktree_id else {
-            return;
-        };
-        let archived = self
-            .state
-            .find_worktree(worktree_id)
-            .map(|(_, w)| w.is_archived)
-            .unwrap_or(false);
-        self.state.set_archived(worktree_id, !archived);
-        self.rebuild_rows();
-        self.sync_sidebar_index();
-        self.reload_detail();
-    }
-
-    fn action_delete(&mut self) {
-        if let Some((_repo, worktree)) = self.state.selected_worktree() {
-            let name = worktree.name.clone();
-            let id = worktree.id;
-            self.modals.push(Modal::Confirm(ConfirmModal::new(
-                "Delete Worktree",
-                format!("Permanently delete '{name}'?"),
-                ConfirmAction::DeleteWorktree(id),
-            )));
-        } else if let Some(repo) = self.state.selected_repository() {
-            let name = repo.name.clone();
-            let id = repo.id;
-            self.modals.push(Modal::Confirm(ConfirmModal::new(
-                "Remove Repository",
-                format!("Remove '{name}' from forestui?"),
-                ConfirmAction::RemoveRepository(id),
-            )));
-        }
-    }
-
-    fn custom_button(&self, index: usize) -> Option<CustomClaudeButton> {
-        self.settings.custom_buttons.get(index).cloned()
-    }
-
-    fn session_id(&self, index: usize) -> Option<String> {
-        self.sessions
-            .as_ref()
-            .and_then(|s| s.get(index))
-            .map(|s| s.id.clone())
-    }
-
-    pub fn run_action(&mut self, action: Action) {
-        let Some(path) = self.state.selected_path() else {
-            return;
-        };
-
-        // Opening a window in a directory that is gone silently lands the shell
-        // in $HOME and lies about where it is, so refuse instead.
-        let needs_directory = matches!(
-            action,
-            Action::Sync
-                | Action::Editor
-                | Action::Terminal
-                | Action::Files
-                | Action::ClaudeNew
-                | Action::ClaudeYolo
-                | Action::ClaudeCustom(_)
-                | Action::ResumeSession(_)
-                | Action::ResumeYolo(_)
-                | Action::ResumeCustom { .. }
-        );
-        if needs_directory && !crate::util::path_exists(&path) {
-            self.notify(
-                format!("Directory no longer exists: {path}"),
-                Severity::Error,
-            );
-            return;
-        }
-
-        match action {
-            Action::Sync => self.sync(path),
-            Action::AddWorktree => self.action_add_worktree(),
-            Action::Editor => self.open_in_editor(&path),
-            Action::Terminal => self.open_in_terminal(&path),
-            Action::Files => self.open_in_file_manager(&path),
-            Action::ClaudeNew => self.start_claude(&path, None, false, None),
-            Action::ClaudeYolo => self.start_claude(&path, None, true, None),
-            Action::ClaudeCustom(index) => {
-                if let Some(button) = self.custom_button(index) {
-                    self.start_claude(&path, None, false, Some(button));
-                }
-            }
-            Action::ResumeSession(index) => {
-                if let Some(id) = self.session_id(index) {
-                    self.start_claude(&path, Some(&id), false, None);
-                }
-            }
-            Action::ResumeYolo(index) => {
-                if let Some(id) = self.session_id(index) {
-                    self.start_claude(&path, Some(&id), true, None);
-                }
-            }
-            Action::ResumeCustom { button, session } => {
-                if let (Some(id), Some(button)) =
-                    (self.session_id(session), self.custom_button(button))
-                {
-                    self.start_claude(&path, Some(&id), false, Some(button));
-                }
-            }
-            Action::RefreshIssues => {
-                github::invalidate_cache(None);
-                self.issues = None;
-                self.fetch_issues();
-            }
-            Action::CreateFromIssue(index) => {
-                if let Some(repo_id) = self.state.selection.repository_id {
-                    self.pending_issue = Some(index);
-                    self.load_branches_then(repo_id, BranchTarget::CreateFromIssue);
-                }
-            }
-            Action::RemoveRepository | Action::Delete => self.action_delete(),
-            Action::Archive | Action::Unarchive => self.action_toggle_archive(),
-        }
-    }
-
-    fn sync(&mut self, path: String) {
-        self.notify("Syncing...", Severity::Information);
-        let tx = self.tx.clone();
-        tokio::spawn(async move {
-            match git::pull(&path).await {
-                Ok(()) => {
-                    tx.info("Sync complete");
-                    tx.send(AppEvent::ReloadDetail);
-                }
-                Err(error) => tx.error(format!("Sync failed: {error}")),
-            }
-        });
-    }
-
-    fn open_in_editor(&mut self, path: &str) {
-        let editor = self.settings.default_editor.clone();
-
-        if tmux::is_inside_tmux() && tmux::is_tui_editor(&editor) {
-            let name = self.state.tmux_window_name(path);
-            if tmux::create_editor_window(&name, path, &editor) {
-                self.notify(
-                    format!("Opened {editor} in edit:{name}"),
-                    Severity::Information,
-                );
-                return;
-            }
-        }
-
-        // GUI editor, or not inside tmux: spawn it detached.
-        let mut parts = editor.split_whitespace();
-        let Some(program) = parts.next() else {
-            return;
-        };
-        let args: Vec<&str> = parts.collect();
-        match std::process::Command::new(program)
-            .args(&args)
-            .arg(path)
-            .stdout(std::process::Stdio::null())
-            .stderr(std::process::Stdio::null())
-            .spawn()
-        {
-            Ok(_) => self.notify(format!("Opened in {program}"), Severity::Information),
-            Err(_) => self.notify(format!("Editor '{editor}' not found"), Severity::Error),
-        }
-    }
-
-    fn open_in_terminal(&mut self, path: &str) {
-        let name = self.state.tmux_window_name(path);
-        if tmux::create_shell_window(&name, path) {
-            self.notify(
-                format!("Opened terminal in term:{name}"),
-                Severity::Information,
-            );
-        } else {
-            self.notify("Failed to create terminal window", Severity::Error);
-        }
-    }
-
-    fn open_in_file_manager(&mut self, path: &str) {
-        let name = self.state.tmux_window_name(path);
-        if tmux::create_mc_window(&name, path) {
-            self.notify(format!("Opened mc in files:{name}"), Severity::Information);
-        } else {
-            self.notify("Failed to create mc window", Severity::Error);
-        }
-    }
-
-    fn start_claude(
-        &mut self,
-        path: &str,
-        resume_session_id: Option<&str>,
-        yolo: bool,
-        custom: Option<CustomClaudeButton>,
-    ) {
-        let name = self.state.tmux_window_name(path);
-        let window = tmux::create_claude_window(
-            &name,
-            path,
-            resume_session_id,
-            yolo,
-            custom.as_ref().map(|b| b.command.as_str()),
-            custom.as_ref().map(|b| b.prefix.as_str()),
-        );
-
-        match window {
-            Some(window_name) => {
-                let verb = if resume_session_id.is_some() {
-                    "Resuming"
-                } else {
-                    "Started"
-                };
-                let label = match (&custom, yolo) {
-                    (Some(button), _) => format!(" ({})", button.label),
-                    (None, true) => " (YOLO)".to_string(),
-                    (None, false) => String::new(),
-                };
-                self.notify(
-                    format!("{verb} Claude{label} in {window_name}"),
-                    Severity::Information,
-                );
-            }
-            None => self.notify("Failed to create Claude window", Severity::Error),
-        }
-    }
-
-    fn submit_rename(&mut self, field: Field) {
-        let Some((repo, worktree)) = self.state.selected_worktree() else {
-            return;
-        };
-        let repo_path = repo.source_path.clone();
-        let worktree_id = worktree.id;
-        let old_name = worktree.name.clone();
-        let old_branch = worktree.branch.clone();
-        let old_path = PathBuf::from(&worktree.path);
-
-        match field {
-            Field::WorktreeName => {
-                let new_name = self.name_input.value().to_string();
-                if new_name.is_empty() || new_name == old_name {
-                    return;
-                }
-                let new_path = old_path
-                    .parent()
-                    .map(|p| p.join(&new_name))
-                    .unwrap_or_else(|| PathBuf::from(&new_name));
-                if new_path.exists() {
-                    self.notify("Path already exists", Severity::Error);
-                    return;
-                }
-                if let Err(error) = std::fs::rename(&old_path, &new_path) {
-                    self.notify(format!("Rename failed: {error}"), Severity::Error);
-                    return;
-                }
-                let new_path_string = new_path.to_string_lossy().to_string();
-                self.state.update_worktree(worktree_id, |w| {
-                    w.name = new_name.clone();
-                    w.path = new_path_string.clone();
-                });
-                claude_session::migrate_sessions(&old_path, &new_path);
-
-                let tx = self.tx.clone();
-                let target = new_path.clone();
-                tokio::spawn(async move {
-                    if let Err(error) = git::repair_worktree(&repo_path, &target).await {
-                        tx.error(format!("Rename failed: {error}"));
-                    }
-                    // State was already updated and saved before the spawn; the
-                    // pane just needs to re-read sessions and meta for the new path.
-                    tx.send(AppEvent::ReloadDetail);
-                });
-                self.rebuild_rows();
-                self.sync_sidebar_index();
-            }
-            Field::BranchName => {
-                let new_branch = self.branch_input.value().to_string();
-                if new_branch.is_empty() || new_branch == old_branch {
-                    return;
-                }
-                let worktree_path = worktree.path.clone();
-                let tx = self.tx.clone();
-                let requested = new_branch.clone();
-                tokio::spawn(async move {
-                    match git::rename_branch(&worktree_path, &old_branch, &requested).await {
-                        Ok(()) => tx.send(AppEvent::ReloadDetail),
-                        Err(error) => tx.error(format!("Branch rename failed: {error}")),
-                    }
-                });
-                self.state
-                    .update_worktree(worktree_id, |w| w.branch = new_branch.clone());
-                self.rebuild_rows();
-            }
-        }
-    }
-
-    fn add_repository(&mut self, path: String, import_worktrees: bool) {
-        let name = Path::new(&path)
-            .file_name()
-            .map(|n| n.to_string_lossy().to_string())
-            .unwrap_or_else(|| path.clone());
-        let repo = Repository::new(name, path.clone());
-        let repo_id = repo.id;
-        self.state.add_repository(repo);
-        self.state.select_repository(repo_id);
-        self.rebuild_rows();
-        self.sync_sidebar_index();
-        self.reload_detail();
-
-        if import_worktrees {
-            self.import_existing_worktrees(repo_id, path);
-        }
-    }
-
-    fn import_existing_worktrees(&mut self, repo_id: Uuid, source_path: String) {
-        let forest_dir = settings_service::get_forest_path();
-        let tx = self.tx.clone();
-
-        tokio::spawn(async move {
-            let listed = match git::list_worktrees(&source_path).await {
-                Ok(listed) => listed,
-                Err(error) => {
-                    tx.error(format!("Failed to import worktrees: {error}"));
-                    return;
-                }
-            };
-
-            let source_resolved = std::fs::canonicalize(&source_path).ok();
-            let mut worktrees = Vec::new();
-
-            for info in &listed {
-                let candidate = std::fs::canonicalize(&info.path).ok();
-                if candidate.is_some() && candidate == source_resolved {
-                    continue;
-                }
-                if info
-                    .path
-                    .starts_with(&forest_dir.to_string_lossy().to_string())
-                {
-                    continue;
-                }
-                let name = Path::new(&info.path)
-                    .file_name()
-                    .map(|n| n.to_string_lossy().to_string())
-                    .unwrap_or_else(|| info.path.clone());
-                let branch = info.branch.clone().unwrap_or_else(|| "HEAD".to_string());
-                worktrees.push(Worktree::new(name, branch, info.path.clone()));
-            }
-
-            tx.info(format!("Imported {} worktrees", worktrees.len()));
-            // The main loop folds these into state and saves; writing the
-            // config file from here would race a save the user's next action
-            // makes in the meantime.
-            tx.send(AppEvent::WorktreesImported { repo_id, worktrees });
-        });
-    }
-
-    fn create_worktree(
-        &mut self,
-        repo_id: Uuid,
-        name: String,
-        branch: String,
-        new_branch: bool,
-        base_branch: Option<String>,
-        pull_first: bool,
-    ) {
-        let Some(repo) = self.state.find_repository(repo_id) else {
-            return;
-        };
-        let source_path = repo.source_path.clone();
-        let worktree_path = settings_service::get_forest_path()
-            .join(&repo.name)
-            .join(&name);
-        let tx = self.tx.clone();
-
-        tokio::spawn(async move {
-            if pull_first {
-                tx.info("Pulling repo...");
-                let _ = git::pull(&source_path).await;
-            }
-
-            // Record where the worktree stemmed from, for the detail header.
-            let (base, base_ref) = if let Some(base) = base_branch.clone() {
-                let reference = git::get_ref(&source_path, &base).await;
-                (Some(base), reference)
-            } else if new_branch {
-                let current = git::get_current_branch(&source_path).await.ok();
-                (current, git::get_ref(&source_path, "HEAD").await)
-            } else {
-                (
-                    Some(branch.clone()),
-                    git::get_ref(&source_path, &branch).await,
-                )
-            };
-
-            if let Err(error) = git::create_worktree(
-                &source_path,
-                &worktree_path,
-                &branch,
-                new_branch,
-                base_branch.as_deref(),
-            )
-            .await
-            {
-                tx.error(format!("Failed to create worktree: {error}"));
-                return;
-            }
-
-            let mut worktree = Worktree::new(
-                name.clone(),
-                branch,
-                worktree_path.to_string_lossy().to_string(),
-            );
-            worktree.base_branch = base;
-            worktree.created_from_ref = base_ref;
-
-            tx.info(format!("Created worktree '{name}'"));
-            tx.send(AppEvent::WorktreeAdded {
-                repo_id,
-                worktree: Box::new(worktree),
-            });
-        });
-    }
 }
 
 #[cfg(test)]
 pub(crate) mod test_support {
     use super::*;
-    use crate::models::Selection;
+    use crate::models::{Repository, Selection, Worktree};
 
     /// An app backed by a throwaway forest directory, with one repository that
     /// has two active worktrees and one archived worktree.
@@ -1811,7 +637,9 @@ pub(crate) mod test_support {
 mod tests {
     use super::test_support::{app_with_fixture, key};
     use super::*;
+    use crate::models::{CustomClaudeButton, Repository, Worktree};
     use ratatui::crossterm::event::KeyCode;
+    use std::path::PathBuf;
 
     #[tokio::test]
     async fn sidebar_lists_repository_then_active_worktrees() {
